+++ src/lib/components/example-collection.ts
@@ -15,7 +15,6 @@ import { TerraTextInputComponentExample } from './forms/input/text-input/example
 import { TerraCheckboxComponentExample } from './forms/checkbox/example/terra-checkbox.component.example';
 import { TerraMultiCheckBoxComponentExample } from './forms/multi-check-box/example/terra-multi-check-box.component.example';
 import { TerraSelectBoxComponentExample } from './forms/select-box/example/terra-select-box.component.example';
-import { TerraSliderComponentExample } from './forms/slider/example/terra-slider.component.example';
 import { TerraSuggestionBoxComponentExample } from './forms/suggestion-box/example/terra-suggestion-box.component.example';
 import { TerraDataTableComponentExample } from './tables/data-table/example/terra-data-table.component.example';
 import { TerraTagComponentExample } from './layouts/tag/example/terra-tag.component.example';
@@ -44,7 +43,6 @@ import { TerraFormsExampleComponent } from './forms/example/terra-forms.example'
 import { RadioGroupComponentExample } from './forms/input/radio/example/radio-group.component.example';
 import { DataTableTemplateComponentExample } from './tables/data-table/example/data-table-template.component.example';
 import { FilterComponentExample } from './filter/example/filter.component.example';
-import { CheckboxGroupComponentExample } from './forms/checkbox-group/example/checkbox-group.component.example';
 import { TerraFormComponentExample } from './forms/form/example/terra-form.component.example';
 import { CkEditorDirectiveExample } from './editors/ck-editor/example/ck-editor-directive-example.component';
 import { TerraFileBrowserComponentExample } from './file-browser/example/terra-file-browser.component.example';
@@ -69,7 +67,6 @@ export const examples: Array<Type<any>> = [
     TerraCheckboxComponentExample,
     TerraMultiCheckBoxComponentExample,
     TerraSelectBoxComponentExample,
-    TerraSliderComponentExample,
     TerraSuggestionBoxComponentExample,
     TerraDataTableComponentExample,
     TerraTagComponentExample,
@@ -98,7 +95,6 @@ export const examples: Array<Type<any>> = [
     RadioGroupComponentExample,
     DataTableTemplateComponentExample,
     FilterComponentExample,
-    CheckboxGroupComponentExample,
     TerraFormComponentExample,
     CkEditorDirectiveExample,
     TerraFileBrowserComponentExample,